requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.3
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyodbc==5.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.2
wheel==0.44.0
wrapt==1.16.0
